requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.9
django-ipware==7.0.1
django-session-timeout==0.1.0
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.2.0
python-decouple==3.8
python-ipware==3.0.0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.1
